requirements.txt
faster-whisper
pyannote.audio
sentence-transformers
ruptures
torch
torchaudio
